Moved helpers to a separate file
